+++ src/depack.js
@@ -1,8 +1,0 @@
-/* license-copyright */
-
-import '../types/externs'
-import myNewPackage from './'
-
-module.exports = {
-  '_myNewPackage': myNewPackage,
-}