+++ src/stdlib.js
@@ -1,14 +1,5 @@
 import { c } from 'erte'
-/* bin-start */
-import argufy, { reduceUsage } from 'argufy'
-import usually from 'usually'
-import indicatrix from 'indicatrix'
-/* bin-end */
 
 module.exports = {
-  'c': c,/* bin-start */
-  'reduceUsage': reduceUsage,
-  'argufy': argufy,
-  'usually': usually,
-  'indicatrix': indicatrix,/* bin-end */
+  'c': c,
 }