+++ test/context/index.js
@@ -2,7 +2,7 @@ import { join } from 'path'
 import { debuglog } from 'util'
 import { readFileSync } from 'fs'
 
-const LOG = debuglog('my-new-package')
+const LOG = debuglog('@artdeco/typescript')
 
 /**
  * A testing context for the package.
@@ -41,24 +41,4 @@ export default class Context {
   async _destroy() {
     LOG('destroy context')
   }
-  static get BIN() {
-    return BIN
-  }
 }
-
-let BIN = 'src/BIN'
-if (process.env.ALAMODE_ENV == 'test-build') {
-  console.log('Testing build bin...')
-  BIN = 'build/bin/mnp'
-} else if (process.env.ALAMODE_ENV == 'test-compile') {
-  console.log('Testing compile bin...')
-  BIN = 'compile/bin/mnp'
-}
-
-/** @typedef {Object<string, Test & TestSuite4>} TestSuite */
-/** @typedef {Object<string, Test & TestSuite3>} TestSuite4 */
-/** @typedef {Object<string, Test & TestSuite2>} TestSuite3 */
-/** @typedef {Object<string, Test & TestSuite1>} TestSuite2 */
-/** @typedef {Object<string, Test & TestSuite0>} TestSuite1 */
-/** @typedef {Object<string, Test>} TestSuite0 */
-/** @typedef {(c: Context)} Test */